Safari-611.1.10/LayoutTests/accessibility/table-with-mismatch-column-count-in-initial-section-expected.txt

Test
Cell 1	Cell 2
This tests that if a table's initial section has fewer columns than the rest of the table, we'll still detect the total column count correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table.columnCount is 2
PASS cell.isEqual(accessibilityController.accessibleElementById('cell2')) is true
PASS successfullyParsed is true

TEST COMPLETE
